+++ OzzCodeGen/CodeEngines/CsSqliteRepository/CSharpSqliteRepositoryEngine.cs
@@ -16,7 +16,7 @@ using System.Xml.Serialization;
 namespace OzzCodeGen.CodeEngines.CsSqliteRepository;
 
 [XmlInclude(typeof(SqliteRepositoryEntitySetting))]
-public class CSharpSqliteRepositoryEngine : BaseCodeEngine
+public class CSharpSqliteRepositoryEngine : BaseAppInfraCodeEngine
 {
     public override string EngineId => EngineTypes.CSharpSqliteRepositoryEngineId;
 
@@ -184,6 +184,23 @@ public class CSharpSqliteRepositoryEngine : BaseCodeEngine
             if (_baseRepositoryClassName == value) return;
             _baseRepositoryClassName = value;
             RaisePropertyChanged(nameof(BaseRepositoryClassName));
+        }
+    }
+
+    [XmlIgnore]
+    [JsonIgnore]
+    public override string TargetInfrastructureDirectory
+    {
+        get
+        {
+            if (string.IsNullOrWhiteSpace(InfrastructureFolder))
+            {
+                return $"If this is empty or whitespace, generated base/contracts will be placed in the target folder,\r\n{TargetDirectory}";
+            }
+            else
+            {
+                return Path.GetFullPath(Path.Combine(TargetDirectory, InfrastructureFolder));
+            }
         }
     }
     private string _baseRepositoryClassName;